requirements.txt
python-can==4.1.0
pyusb==1.2.1
pyserial==3.5
msgpack==1.0.4
python-ics==910.10.post6
canalystii==0.1
gs_usb==0.3.0
#cantact==0.1.2
nixnet==0.3.2
